feat: 사용자 결제 요청 테이블 생성, 테이블 매핑 관계 수정
- UserPaymentOrder
- Payment: [refactor]연관관계 매핑 수정

#68

+++ src/main/java/com/i5e2/likeawesomevegetable/domain/payment/api/Payment.java
@@ -23,7 +23,7 @@ public class Payment {
     private UserPaymentOrder userPaymentOrder;
 
     @Column(name = "payment_amount")
-    private Integer paymentAmount;
+    private Long paymentAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "payment_method")
